feat: impl task apis

+++ src/main/java/com/example/task_management/services/TaskService.java
@@ -24,52 +24,94 @@ public class TaskService {
         this.taskListRepository = taskListRepository;
     }
 
-    public List<Task> getAllTasks() {
+    public List<Task> getAll() {
         var allTaskRecords = taskRepository.findAll();
         return allTaskRecords.stream()
-            .map(Task::new)
-            .toList();
+                .map(Task::new)
+                .toList();
     }
 
-    public Task getTaskById(BigInteger id) {
+    public Task getById(BigInteger id) {
         return taskRepository.findById(id)
-            .map(Task::new)
-            .orElseThrow(() -> new EntityNotFoundException(MessageFormat.format("Task with id {0} not found", id)));
+                .map(Task::new)
+                .orElseThrow(
+                        () -> new EntityNotFoundException(
+                                MessageFormat.format("Task with id {0} not found", id)));
     }
 
-    public Task createTask(String name, String description, BigInteger parentTaskListId) {
-        var parentTaskList = taskListRepository.findById(parentTaskListId); // Fetch TaskListRecord by parentTaskListId
-        if (parentTaskList.isPresent()) {
+    public List<Task> getAllByParentId(BigInteger taskListId) {
+        var parentTaskList = taskListRepository.findById(taskListId);
+        if (parentTaskList.isEmpty()) {
+            throw new EntityNotFoundException(
+                    MessageFormat.format("Task list with id {0} not found when getting tasks", taskListId));
+        } else {
+            var taskListRecords = taskRepository.findAllByParentTaskListId(taskListId);
+            return taskListRecords.stream()
+                .map(Task::new)
+                .toList();
+        }
+    }
+
+    public Task create(String name, String description, BigInteger parentTaskListId) {
+        var parentTaskList = taskListRepository.findById(parentTaskListId);
+        if (parentTaskList.isEmpty()) {
+            throw new EntityNotFoundException(
+                    MessageFormat.format("Task list with id {0} not found when creating a new task", parentTaskListId));
+        } else {
             var taskRecord = new TaskRecord();
             taskRecord.setName(name);
             taskRecord.setDescription(description);
             taskRecord.setParentTaskList(parentTaskList.get());
+
             TaskRecord newTaskRecord = taskRepository.save(taskRecord);
             return new Task(newTaskRecord);
-        } else {
-            throw new EntityNotFoundException(MessageFormat.format("TaskList with id {0} not found", parentTaskListId));
         }
     }
 
-    public Task updateTask(BigInteger id, String name, String description) {
-        var taskToBeUpdated = taskRepository.findById(id);
-        if (taskToBeUpdated.isPresent()) {
-            var taskRecord = taskToBeUpdated.get();
-            taskRecord.setName(name);
-            taskRecord.setDescription(description);
-            TaskRecord updatedTaskRecord = taskRepository.save(taskRecord);
-            return new Task(updatedTaskRecord);
+    public Task update(BigInteger id, String name, String description, BigInteger parentTaskListId, String ownerEmail) {
+        System.out.println("Parent Task List ID: " + parentTaskListId);
+        var taskRecordOpt = taskRepository.findById(id);
+        if (taskRecordOpt.isPresent()) {
+            var taskRecord = taskRecordOpt.get();
+            if (isCurrentUserOwnThisResource(taskRecord, ownerEmail)) {
+                var newName = name != null ? name : taskRecord.getName();
+                var newDescription = description != null ? description : taskRecord.getDescription();
+                var newTaskListParent = parentTaskListId != null
+                        ? taskListRepository.findById(parentTaskListId)
+                                .orElseThrow(() -> new EntityNotFoundException(
+                                        MessageFormat.format("Task list with id {0} not found when updating task list", parentTaskListId)))
+                        : taskRecord.getParentTaskList();
+                
+                taskRecord.setName(newName);
+                taskRecord.setDescription(newDescription);
+                taskRecord.setParentTaskList(newTaskListParent);
+                var updatedTaskRecord = taskRepository.save(taskRecord);
+                return new Task(updatedTaskRecord);
+            } else {
+                throw new SecurityException("You do not have permission to update this task");
+            }
         } else {
             throw new EntityNotFoundException(MessageFormat.format("Task with id {0} not found", id));
         }
     }
 
-    public void deleteTask(BigInteger id) {
+    public void delete(BigInteger id, String ownerEmail) {
         var taskTobeDeleted = taskRepository.findById(id);
         if (taskTobeDeleted.isPresent()) {
-            taskRepository.deleteById(id);
+            if (isCurrentUserOwnThisResource(taskTobeDeleted.get(), ownerEmail)) {
+                taskRepository.deleteById(id);
+            } else {
+                throw new SecurityException("You do not have permission to delete this task");
+            }
         } else {
             throw new EntityNotFoundException(MessageFormat.format("Task with id {0} not found", id));
         }
+    }
+
+    private boolean isCurrentUserOwnThisResource(TaskRecord taskRecord, String ownerEmail) {
+        return taskRecord
+            .getParentTaskList()
+            .getParentBoard()
+            .getParentWorkspace().getOwnerEmail().equals(ownerEmail);
     }
 }
